kernell + details seller + crud seller profile

+++ database/migrations/2023_09_15_210935_add_verification_field_to_freelancer_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::table('freelancer', function (Blueprint $table) {
             $table->string('link')->nullable();
-            $table->unsignedInteger('id_card')->nullable();
-            $table->unsignedInteger('id_card_with_selfie')->nullable();
+            $table->unsignedBigInteger('id_card')->nullable();
+            $table->unsignedBigInteger('id_card_with_selfie')->nullable();
         });
     }
